requirements.txt
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
Django==4.0.4
django-browser-reload==1.5.0
django-ckeditor==6.4.1
django-js-asset==2.0.0
django-tailwind==3.1.1
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==6.1.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
